requirements.txt
web.py==0.38
nltk==3.2.4
dill
gunicorn
pandas
numpy
scipy
scikit-learn
sklearn-crfsuite